// front-end/job-portal/src/app/model/job-application.model.ts
export interface JobApplication {
    id?: number;               
    applicantName: string;      
    applicantEmail: string;     
    applicantPhone: string;     
    resumeUrl: string;         
    experience: number;         
    minSalary: number;         
    maxSalary: number;         
    coverLetter: string;       
    jobId: number;              
  }
  